Moved test to existing file
