+++ clang/test/CodeGenCXX/control-flow-in-stmt-expr.cpp
@@ -1,4 +1,5 @@
-// RUN: %clang_cc1 --std=c++20 -triple x86_64-linux-gnu -emit-llvm %s -o - | FileCheck %s
+// RUN: %clang_cc1 --std=c++20 -fexceptions -triple x86_64-linux-gnu -emit-llvm %s -o - | FileCheck -check-prefixes=EH %s
+// RUN: %clang_cc1 --std=c++20 -triple x86_64-linux-gnu -emit-llvm %s -o - | FileCheck -check-prefixes=NOEH,CHECK %s
 
 struct Printy {
   Printy(const char *name) : name(name) {}
@@ -347,6 +348,59 @@ void NewArrayInit() {
 
   // CHECK:       arraydestroy.done{{.*}}:                               ; preds = %arraydestroy.body, %if.then
   // CHECK-NEXT:    br label %return
+}
+
+namespace std {
+  typedef decltype(sizeof(int)) size_t;
+
+  // libc++'s implementation
+  template <class _E>
+  class initializer_list
+  {
+    const _E* __begin_;
+    size_t    __size_;
+    initializer_list(const _E* __b, size_t __s);
+  };
+}
+
+class Object {
+public:
+  Object() = default;
+  struct KV;
+  Object(std::initializer_list<KV>);
+  Object(KV);
+};
+
+class Value {
+public:
+  Value(std::initializer_list<Value>);
+  Value(const char *V);
+  ~Value();
+};
+
+class ObjectKey {
+public:
+  ObjectKey(const char *S);
+  ~ObjectKey();
+};
+
+struct Object::KV {
+  ObjectKey K;
+  Value V;
+};
+
+void DestroyInConditionalCleanup() {
+  // EH-LABEL: DestroyInConditionalCleanupv()
+  // NOEH-LABEL: DestroyInConditionalCleanupv()
+  // Verify we use conditional cleanups.
+  foo() ? Object{{"key1", {"val1", "val2"}}} : Object{{"key2", "val2"}};
+  // NOEH:   cond.true:
+  // NOEH:      call void @_ZN9ObjectKeyC1EPKc
+  // NOEH:      store ptr %K, ptr %cond-cleanup.save
+
+  // EH:   cond.true:
+  // EH:        invoke void @_ZN9ObjectKeyC1EPKc
+  // EH:        store ptr %K, ptr %cond-cleanup.save
 }
 
 void ArrayInitWithContinue() {
